Moved AppConfig to utils.application
